No longer outputs initial values, as they are really unreliable, and prevent a consumer knowing if a control has actually ever been signalled.
Control now has a Timestamp property to indicate when it was last changed.  This makes it easy to see how 'stale' data is, as well as assess latency (after initial warm up, it currently quickly settles to <0.1ms to notify the subscriber or a batch of changes.

Controllers are now 'zombied' meaning that the same instance will be associated with a Controller's container even when unplugged and plugged back in (unless the controller's description report changes).  This should make it much easier for any consumer that holds onto a reference to a Controller, or it's Controls.

Improved the Sample system and added publish profiles.