Improved extension methods by extending interface and adding new Connected() extension method that only returns connected Devices.
Device also improved so that monitoring the ConnectionState actually attempts to connect to the device, previously no change would occur when subscribing to this observable, unless you also subscribed to the Device itself.

Also moved logic for UsagesAll and UsagesAny (plus ControlUsagesAll and ControlUsagesAny) into their own extension methods that extend Device, making it possible to use them in the Connected() and Connect() methods directly.